backend/requirements.txt
Flask==3.0.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
xgboost==1.7.6
scikit-learn==1.3.0
gunicorn==21.2.0
